programa auto resuelto
Zoe Caballero Dominguez
A01747247
grupo 04

+++ auto.py
@@ -3,9 +3,9 @@
 
 # Escribe tu programa después de esta línea.
 velocidad = int(input("Velocidad del auto en km/h:"))
-primeraDistancia = float(velocidad * 7)
-segundaDistancia = float(velocidad * 4.5)
-tiempo = float(791/velocidad)
+primeraDistancia = velocidad * 7
+segundaDistancia = velocidad * 4.5
+tiempo = 791/velocidad
 print ("Distancia recorrida en 7 hrs: %.1f km" % (primeraDistancia))
 print ("Distancia recorrida en 4.5 hrs: %.1f km" % (segundaDistancia))
 print ("Tiempo para recorrer 791 km: %.1f hrs" % (tiempo))
